kernels: RVV + Gemmini optimization campaign on AWS F2
Hardware-measured on f2_dual_small_norose_tacit_q31_60mhz. Speedups are
against each backend's own original curated kernels, not against scalar.

  dronet      rvv        17,813,964 -> 7,496,398   2.38x  err=0
  yolov8_nano rvv       333,175,953 -> 165,326,342 2.02x  err=0
  fused_full  pure-fp16  60,371,259 -> 7,710,288   7.83x
  fused_full  int8-hyb    7,587,275 -> 4,922,762   1.54x
  dronet      gemmini    20,933,650 -> 18,311,704  1.14x  err=0

Kernel work
  conv2d_s8: pixel register tiling (TILE=4 cascade 4->2->1); the OW%4
    behaviour was predicted a priori and confirmed 16/16 across two models.
  conv2d_f16: real L1D cache-blocking outer loop + vfwmacc.vf replacing
    vfwcvt+vfmacc. 3.08x, bit-identical output.
  maxpool2d_s8: vlse8 stride-2 gather -> contiguous vle16 + vnsrl
    deinterleave, one load serving two taps. 3.44x.
  silu_s8: vluxei8 memory gather -> vrgather.vv over a register-resident
    256B LUT. 3.10x. Memory-indexed gather measures ~11.9 cyc/elem against
    ~1.7 for an arithmetic chain.
  batchnorm2d_s8, add_s8, cast_i8_to_f16, linear_f16, lstm_f16: fold
    loop-invariant scales; N- and gate-blocking by 4 to match fmaPipeDepth.
  cat2/3/4 on gemmini: MVIN_SCALE + tiled_matmul_auto(full_C=true) drains
    the raw accumulator, bypassing the inexact mvout scale unit. err=0.

Correctness
  conv2d_s8 indexed tensors with 32-bit arithmetic while its sibling used
    size_t; BSS lives above 0x80000000 so the index wrapped. Presented as
    err=1/3/11 or a Store/AMO fault depending on binary layout, and was
    invisible to spike. Two of three siblings already had the fix.
  gemmini_tiled_conv_pool and gemmini_resadd_relu declared bit_exact and
    measured err=17 in isolation; archived with the cause documented.
  gemmini_q31_add_s8: hardware resadd branch removed after proving no exact
    construction exists (every resadd entry point hardcodes full_C=false).
    err=43 -> 0 on yolov8n, the model that actually exercises the gate.
  gemmini_q31 driven to err=0 via im2col_full_C's exact two-stage
    requantize; tiled_conv keeps one hardware round-shift and cannot match
    the golden's double rounding on this datapath.

Pipeline
  extract_graph: graph-level BN folding before quantization (57/57 yolov8n,
    3/6 dronet -- pre-activation blocks are correctly left alone).
    Accuracy verified end-to-end; per-channel weight quant not required.
  generate_skeleton: weight-layout agreement trip-wire. A shared
    target_affinity edit silently transposed an unrelated op's weights and
    surfaced as uniform error in untouched kernels; the build now refuses.
  generate_kernels: distinguish a BUILD failure from a numeric mismatch in
    curated verify. One non-compiling kernel was downgrading every op
    sorting before it (worth 21% of yolov8n).
  roofline_analysis: dtype-aware peaks. int8 is 8 ops/cyc (vwmacc only, no
    Zvqdotq), fp16 16 narrow / 8 widened -- not the dtype-blind 64. Also
    fp16 formulas sat after the __main__ guard and had never been reached
    by any CLI run.

+++ kernels/rvv/rvv_linear_s8_pc_direct.c
@@ -22,8 +22,8 @@ void kernel_linear_s8_pc(const int8_t *input, const int8_t *weight,
         for (int n = 0; n < N; n++) {
             size_t vl;
             vint32m4_t vacc = __riscv_vmv_v_x_i32m4(0, __riscv_vsetvlmax_e32m4());
-            const int8_t *in_row = input + m * K;
-            const int8_t *w_row  = weight + n * K;
+            const int8_t *in_row = input + (size_t)m * K;
+            const int8_t *w_row  = weight + (size_t)n * K;
             int k = 0;
             for (; k + 2 * (int)(__riscv_vsetvlmax_e8m1()) <= K;
                    k += 2 * (int)(__riscv_vsetvlmax_e8m1())) {
@@ -64,7 +64,7 @@ void kernel_linear_s8_pc(const int8_t *input, const int8_t *weight,
             scaled += output_offset;
             if (scaled < activation_min) scaled = activation_min;
             if (scaled > activation_max) scaled = activation_max;
-            output[m * N + n] = (int8_t)scaled;
+            output[(size_t)m * N + n] = (int8_t)scaled;
         }
     }
 }
